--- Tangem/App/Services/EnvironmentStorage/EnvironmentStorageKeys.swift
@@ -1,14 +1,0 @@
-//
-//  EnvironmentStorageKeys.swift
-//  Tangem
-//
-//  Created by Sergey Balashov on 26.10.2022.
-//  Copyright © 2022 Tangem AG. All rights reserved.
-//
-
-import Foundation
-
-enum EnvironmentStorageKeys: String {
-    case testnet = "testnet"
-    case integratedFeatures = "integrated_features"
-}
